refactor: renamed rai_sim and rai_bench
