+++ pyproject.toml
@@ -17,8 +17,8 @@ python = "^3.10, <3.13"
 rai = {path = "src/rai_core", develop = true}
 rai_asr = {path = "src/rai_asr", develop = true}
 rai_tts = {path = "src/rai_tts", develop = true}
-rai_benchmarks = {path = "src/rai_benchmarks", develop = true}
-rai_simulations = {path = "src/rai_simulations", develop = true}
+rai_bench = {path = "src/rai_bench", develop = true}
+rai_sim = {path = "src/rai_sim", develop = true}
 
 langchain-core = "^0.3"
 langchain = "*"
