--- src/rai_simulations/rai_simulations/example_scene2.yaml
@@ -1,26 +1,0 @@
-binary_path: /home/mkotynia/repos/RAI/RAIManipulationDemoBenchmark_jammyhumble/RAIManipulationDemo/RAIManipulationDemo.GameLauncher # TODO (mk) remove this before finalization
-entities:
-  - name: cube1
-    prefab_name: carrot
-    pose:
-      translation:
-        x: 0.5
-        y: -0.3
-        z: 0.05
-      rotation:
-        x: 0.0
-        y: 0.0
-        z: 0.0
-        w: 1.0
-  - name: cube2
-    prefab_name: corn
-    pose:
-      translation:
-        x: 0.5
-        y: 0.3
-        z: 0.05
-      rotation:
-        x: 0.0
-        y: 0.0
-        z: 0.0
-        w: 1.0
